requirements.txt
pandas~=2.2.3
matplotlib~=3.10.0
python-telegram-bot~=21.10
transformers~=4.48.2